feat: improve wallet Identity
