+++ apps/frontend/src/app/5_pages/App/App.tsx
@@ -92,7 +92,9 @@ function App() {
           onDisconnect={() => {}}
           hideSubmenu
         />
-        {/* <EthersProviderTest /> */}
+        <br />
+        <br />
+        <EthersProviderTest />
       </main>
     </div>
   );
